MSIM: Fix no service notification for manual search mode.
If SIM card is deactivated and network selection is chosen
as manual search mode, an incorrect indication 'No service'
shown to user.Here the No service is shown as a result of
Out of Service state caused by sim deactivation.

Fix: Check for Out of Service state caused by SIM deactivation
and dont show user the no service indication for this case.

Change-Id: I18c03bde425069e21ff16238d72761efba111993
CRs-Fixed: 1021877.